--- program_management/views/group_element_year/__init__.py
@@ -1,1 +1,0 @@
-from . import update, read, delete, common, create
