+++ program_management/views/groupelementyear_delete.py
@@ -31,8 +31,8 @@ from django.views.generic import DeleteView
 from base.views.common import display_error_messages, display_success_messages, display_warning_messages
 from program_management.business.group_element_years.detach import DetachEducationGroupYearStrategy, \
     DetachLearningUnitYearStrategy
-from program_management.views.group_element_year import perms as group_element_year_perms
-from program_management.views.group_element_year.common import GenericGroupElementYearMixin
+from program_management.views import perms as group_element_year_perms
+from program_management.views.generic import GenericGroupElementYearMixin
 
 
 class DetachGroupElementYearView(GenericGroupElementYearMixin, DeleteView):
