+++ program_management/views/prerequisite_read.py
@@ -6,7 +6,7 @@
 #    The core business involves the administration of students, teachers,
 #    courses, programs and so on.
 #
-#    Copyright (C) 2015-2019 Université catholique de Louvain (http://www.uclouvain.be)
+#    Copyright (C) 2015-2018 Université catholique de Louvain (http://www.uclouvain.be)
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -23,84 +23,19 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-import json
-
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.db.models import Prefetch
-from django.shortcuts import get_object_or_404
-from django.utils.decorators import method_decorator
 from django.utils.translation import gettext_lazy as _
-from django.views.generic import DetailView
 
 from base.business.education_groups import perms
 from base.models import group_element_year
 from base.models.education_group_year import EducationGroupYear
 from base.models.enums.education_group_categories import Categories
-from base.models.enums.education_group_types import TrainingType, MiniTrainingType, GroupType
-from base.models.group_element_year import find_learning_unit_formations
-from base.models.learning_unit_year import LearningUnitYear
-from base.models.person import Person
 from base.models.prerequisite import Prerequisite
 from base.models.utils.utils import get_object_or_none
 from base.views.common import display_warning_messages
-from base.views.education_groups.detail import CatalogGenericDetailView
-from program_management.business.group_element_years.group_element_year_tree import EducationGroupHierarchy
 from program_management.business.learning_units.prerequisite import \
     get_prerequisite_acronyms_which_are_outside_of_education_group
-
-NO_PREREQUISITES = TrainingType.finality_types() + [
-    MiniTrainingType.OPTION.name,
-    MiniTrainingType.MOBILITY_PARTNERSHIP.name,
-] + GroupType.get_names()
-
-
-@method_decorator(login_required, name='dispatch')
-class LearningUnitGenericDetailView(PermissionRequiredMixin, DetailView, CatalogGenericDetailView):
-    model = LearningUnitYear
-    context_object_name = "learning_unit_year"
-    pk_url_kwarg = 'learning_unit_year_id'
-
-    permission_required = 'base.can_access_education_group'
-    raise_exception = True
-
-    def get_person(self):
-        return get_object_or_404(Person, user=self.request.user)
-
-    def get_root(self):
-        return get_object_or_404(EducationGroupYear, pk=self.kwargs.get("root_id"))
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        root = self.get_root()
-        self.hierarchy = EducationGroupHierarchy(root, tab_to_show=self.request.GET.get("tab_to_show"))
-        # TODO remove parent in context
-        context['person'] = self.get_person()
-        context['root'] = root
-        context['root_id'] = root.pk
-        context['parent'] = root
-        context['tree'] = json.dumps(self.hierarchy.to_json())
-        context['group_to_parent'] = self.request.GET.get("group_to_parent") or '0'
-        context['show_prerequisites'] = self.show_prerequisites(root)
-        context['selected_element_clipboard'] = self.get_selected_element_for_clipboard()
-        return context
-
-    def show_prerequisites(self, education_group_year):
-        return education_group_year.education_group_type.name not in NO_PREREQUISITES
-
-
-class LearningUnitUtilization(LearningUnitGenericDetailView):
-    template_name = "learning_unit/tab_utilization.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["group_element_years"] = self.object.child_leaf.select_related("parent")
-        context["formations"] = find_learning_unit_formations(
-            list(grp.parent for grp in self.object.child_leaf.select_related("parent")),
-            parents_as_instances=True
-        )
-        return context
+from program_management.views.generic import LearningUnitGenericDetailView
 
 
 class LearningUnitPrerequisite(LearningUnitGenericDetailView):
